// medical-clinics-frontend/src/app/services/models/work-day-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface WorkDayDto {

  /**
   * case insensitive
   */
  dayName: 'MONDAY' | 'TUESDAY' | 'WEDNESDAY' | 'THURSDAY' | 'FRIDAY' | 'SATURDAY' | 'SUNDAY';
  endOfWorkingDay: string;
  startOfWorkingDay: string;
}